+++ molutils/__init__.py
@@ -2,8 +2,10 @@ from ._version import __version__
 
 from ._parsers.parser_pdb import ParserPDB
 
+from ._ui.app_subcommand import AppSubcommand
+
 from ._core.list_ import List
 from ._core.count import Count
 from ._core.extract import Extract
 
-from ._ui.app import App
+from ._ui.app_main import AppMain