+++ molutils/_core/count.py
@@ -1,8 +1,26 @@
 from pathlib import Path
 
+import molutils as mu
+
 # //////////////////////////////////////////////////////////////////////////////
-class Count:
+class Count(mu.AppSubcommand):
+    # -------------------------------------------------------------------------- UI SECTION
+    def run(self):
+        command = self.main.subcommands.pop(0)
+
+        if command == "models": return self.app_count_models()
+
+        raise ValueError(f"Unknown command: {command}")
+
+
     # --------------------------------------------------------------------------
+    def app_count_models(self):
+        path_in = self.main.get_arg_path("path_in")
+        self.main.assert_file_in(path_in)
+        print(mu.Count.models(path_in))
+
+
+    # -------------------------------------------------------------------------- LOGIC SECTION
     @classmethod
     def models(cls, path_pdb: Path) -> int:
         data = path_pdb.read_text()