+++ molutils/_core/list_.py
@@ -3,8 +3,33 @@ from pathlib import Path
 import molutils as mu
 
 # //////////////////////////////////////////////////////////////////////////////
-class List:
+class List(mu.AppSubcommand):
+    # -------------------------------------------------------------------------- UI SECTION
+    def run(self):
+        command = self.main.subcommands.pop(0)
+
+        if command == "chains": return self.app_list_chains()
+        if command == "resids": return self.app_list_resids()
+
+        raise ValueError(f"Unknown command: {command}")
+
+
     # --------------------------------------------------------------------------
+    def app_list_chains(self):
+        path_in = self.main.get_arg_path("path_in")
+        first_only = self.main.get_arg_bool("first_only")
+        self.main.assert_file_in(path_in)
+        print(*mu.List.chains(path_in, first_only))
+
+
+    # --------------------------------------------------------------------------
+    def app_list_resids(self):
+        path_in = self.main.get_arg_path("path_in")
+        self.main.assert_file_in(path_in)
+        print(*mu.List.resids(path_in))
+
+
+    # -------------------------------------------------------------------------- LOGIC SECTION
     @classmethod
     def chains(cls, path_pdb: Path, first_only: bool = False) -> list[str]:
         pdb = mu.ParserPDB.from_file(path_pdb)