+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
 
+## [0.3.0] - 2025-04-29
+- Added list subcommand `resids`.
+
+
+## [0.2.0] - 2025-04-27
+- Added `select` subcommand.
+    - selects atoms with a custom `MDAnalysis` query.
+- Added `list` subcommand, with `list chains` child.
+- Added subcommand for saving splitted chains.
+
+
 ## [0.1.0] - 2025-04-25
 - Initial upload to PyPI